Subscribers/requirements.txt
fastapi
uvicorn
kafka-python
pymongo
threading